# fairway360/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fairway360 multi-tenant SaaS API. All data is scoped to the authenticated user's club; club_id is always derived from the session, never from the request.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads/demo-request:
    post:
      operationId: createDemoRequest
      tags: [public]
      summary: Submit a demo / sales request from the marketing site
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DemoRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }

  # ---- Members portal -------------------------------------------------------
  /me/dashboard:
    get:
      operationId: getMemberDashboard
      tags: [members]
      responses:
        "200":
          description: Member dashboard bundle
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MemberDashboard" }
        "401": { $ref: "#/components/responses/Error" }
...
  /payments/checkout:
    post:
      operationId: createCheckout
      tags: [members]
      responses:
        "200":
          description: Stripe Checkout session
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CheckoutSession" }
        "400": { $ref: "#/components/responses/Error" }
...
  /events:
    get:
      operationId: listEvents
      tags: [members]
      responses:
        "200":
          description: Club events
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ClubEvent" }
    post:
      operationId: createEvent
      tags: [supervisor]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EventInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ClubEvent" }
  /events/{id}/rsvp:
    post:
      operationId: rsvpEvent
      tags: [members]
      parameters: [{ $ref: "#/components/parameters/IdPath" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RsvpInput" }
      responses:
        "200":
          description: RSVP recorded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
  /tee-times:
    get:
      operationId: listTeeTimes
      tags: [members]
      parameters:
        - name: open
          in: query
          required: false
          schema: { type: boolean }
          description: When true, return only open (bookable) slots
      responses:
        "200":
          description: Tee times
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TeeTime" }
    post:
      operationId: bookTeeTime
      tags: [members]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TeeTimeInput" }
      responses:
        "201":
          description: Booked
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TeeTime" }
...
  /tee-times/{id}:
    delete:
      operationId: cancelTeeTime
      tags: [members]
      parameters: [{ $ref: "#/components/parameters/IdPath" }]
      responses:
        "204": { description: Cancelled }
  /orders:
    get:
      operationId: listOrders
      tags: [fnb]
      parameters:
        - name: status
          in: query
          required: false
          schema: { $ref: "#/components/schemas/OrderStatus" }
      responses:
        "200":
          description: Orders (F&B service board)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Order" }
    post:
      operationId: createOrder
      tags: [members]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrderInput" }
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
...
  /announcements:
    get:
      operationId: listAnnouncements
      tags: [members]
      responses:
        "200":
          description: Announcements
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Announcement" }
    post:
      operationId: createAnnouncement
      tags: [supervisor]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnnouncementInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Announcement" }

  # ---- Department channels (chat) -------------------------------------------
...
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string }
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ProblemDetails" }
  schemas:
...
    Ok:
      type: object
      properties:
        ok: { type: boolean }
      required: [ok]
...
    ProblemDetails:
      type: object
      properties:
        title: { type: string }
        detail: { type: string }
        status: { type: number }
      required: [title]
...
    DemoRequest:
      type: object
      properties:
        name: { type: string }
        clubName: { type: string }
        email: { type: string }
        phone: { type: string }
        businessType: { type: string }
        problem: { type: string }
        volume: { type: string }
      required: [name, email]
...
    CheckoutSession:
      type: object
      properties:
        url: { type: string }
      required: [url]
...
    MemberAccount:
      type: object
      properties:
        name: { type: string }
        memberSince: { type: string }
        tier: { type: string }
        number: { type: string }
        balance: { type: number }
        handicap: { type: number }
        roundsThisYear: { type: number }
      required: [name, tier, number, balance, roundsThisYear]
    UpcomingItem:
      type: object
      properties:
        id: { type: string }
        label: { type: string }
        date: { type: string }
        time: { type: string }
      required: [id, label, date, time]
    MemberDashboard:
      type: object
      properties:
        account: { $ref: "#/components/schemas/MemberAccount" }
        upcoming:
          type: array
          items: { $ref: "#/components/schemas/UpcomingItem" }
        announcements:
          type: array
          items: { $ref: "#/components/schemas/Announcement" }
        events:
          type: array
          items: { $ref: "#/components/schemas/ClubEvent" }
      required: [account, upcoming, announcements, events]
...
    ClubEvent:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        date: { type: string }
        time: { type: string }
        spots: { type: string }
        tag: { type: string }
      required: [id, title, date, time, spots, tag]
    EventInput:
      type: object
      properties:
        title: { type: string }
        startsAt: { type: string }
        tag: { type: string }
        capacity: { type: number }
      required: [title, startsAt, tag]
    RsvpInput:
      type: object
      properties:
        partySize: { type: number }
      required: [partySize]
...
    TeeTime:
      type: object
      properties:
        id: { type: string }
        member: { type: string }
        time: { type: string }
        date: { type: string }
        players: { type: number }
        holes: { type: number }
        status: { type: string }
      required: [id, time, players, holes, status]
    TeeTimeInput:
      type: object
      properties:
        startsAt: { type: string }
        players: { type: number }
        holes: { type: number }
      required: [startsAt, players, holes]
    OrderStatus:
      type: string
      enum: [New, Preparing, Ready, Delivered]
    OrderLine:
      type: object
      properties:
        itemId: { type: string }
        name: { type: string }
        price: { type: number }
        qty: { type: number }
      required: [itemId, name, price, qty]
    Order:
      type: object
      properties:
        id: { type: string }
        groupId: { type: string }
        member: { type: string }
        hole: { type: number }
        cartNumber: { type: string }
        lines:
          type: array
          items: { $ref: "#/components/schemas/OrderLine" }
        note: { type: string }
        status: { $ref: "#/components/schemas/OrderStatus" }
        placedAt: { type: string }
        total: { type: number }
      required: [id, member, lines, note, status, placedAt, total]
    OrderInputLine:
      type: object
      properties:
        itemId: { type: string }
        qty: { type: number }
      required: [itemId, qty]
    OrderInput:
      type: object
      properties:
        groupId: { type: string, description: "Round id, set by staff placing an order on a group's behalf. Omitted for member self-orders." }
        hole: { type: number }
        note: { type: string }
        lines:
          type: array
          items: { $ref: "#/components/schemas/OrderInputLine" }
      required: [lines]
...
    Announcement:
      type: object
      properties:
        id: { type: string }
        tag: { type: string }
        title: { type: string }
        date: { type: string }
      required: [id, title]
    AnnouncementInput:
      type: object
      properties:
        tag: { type: string }
        title: { type: string }
        body: { type: string }
        audience: { type: string }
      required: [title]
...